Created an instruction generator to map opcodes to instrutions. Fixed tests not using the address bus interface properly.

+++ src/Instructions/InstructionFactory.h
@@ -18,4 +18,4 @@ AddressingMode *createAddressingMode(AddressingModeType type);
  * @param opcode 
  * @return Instruction 
  */
-Instruction *createInstruction(uint16_t opcode);
+Instruction *createInstruction(uint8_t opcode);
